+++ packages/ngrx/src/state.ts
@@ -1,20 +1,12 @@
-// import { ensureStateMetadata } from './internals';
-
-export interface ApiEntityState<T> {
-    collection?: T[];
-    entityId?: string;
-    loading?: boolean;
-    loaded?: boolean;
-}
-
-export interface ApiOptions<T> {
-    name: string;
-    route: string;
-    defaults: T;
-}
+import { ensureStateMetadata } from './internals';
+import { ApiOptions } from './symbols';
 
 export function ApiState<T>(options: ApiOptions<T>) {
     return function(target: any) {
-
+        const metadata = ensureStateMetadata(target);
+        
+        metadata.name = options.name;
+        metadata.route = options.route;
+        metadata.defaults = options.defaults;
     };
 }
