+++ packages/ngrx/src/symbols.ts
@@ -1,3 +1,16 @@
+export interface ApiEntityState<T> {
+    collection?: T[];
+    entityId?: string;
+    loading?: boolean;
+    loaded?: boolean;
+}
+
+export interface ApiOptions<T> {
+    name: string;
+    route: string;
+    defaults: T;
+}
+
 export type ApiMethod =
 | 'GET'
 | 'CREATE'
